+++ regulatory_reporting/api_views/__init__.py
@@ -1,6 +1,10 @@
 from .flaw import FlawSRPReportMilestoneViewSet, FlawSRPReportViewSet
 from .srp_milestone import SRPReportMilestoneViewSet
 from .srp_report import SRPReportViewSet
+from .upstream_mappings import (
+    FlawUpstreamMappingDetailView,
+    FlawUpstreamMappingListCreateView,
+)
 from .upstream_notifications import UpstreamNotificationView, UpstreamProjectView
 
 __all__ = [
@@ -10,4 +14,6 @@ __all__ = [
     "SRPReportViewSet",
     "UpstreamNotificationView",
     "UpstreamProjectView",
+    "FlawUpstreamMappingListCreateView",
+    "FlawUpstreamMappingDetailView",
 ]